Script listening for requests from domoticz. Removed server file.
